move TLS type to its own file
